# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Student {
  id: number;
  name: string;
  enrollmentNo: string;
  email: string;
  semester: number;
  branch: string;
}

export interface StudentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  enrollmentNo: string;
  email: string;
  /** @minimum 1 */
  semester: number;
  /** @minLength 1 */
  branch: string;
}

export type AttendanceStatus = typeof AttendanceStatus[keyof typeof AttendanceStatus];


export const AttendanceStatus = {
  Present: 'Present',
  Absent: 'Absent',
} as const;

export interface Attendance {
  id: number;
  studentId: number;
  subject: string;
  date: string;
  status: AttendanceStatus;
}

export type AttendanceInputStatus = typeof AttendanceInputStatus[keyof typeof AttendanceInputStatus];


export const AttendanceInputStatus = {
  Present: 'Present',
  Absent: 'Absent',
} as const;

export interface AttendanceInput {
  studentId: number;
  /** @minLength 1 */
  subject: string;
  date: string;
  status: AttendanceInputStatus;
}

export interface Performance {
  id: number;
  studentId: number;
  subject: string;
  internalMarks: number;
  practicalMarks: number;
}

export interface PerformanceInput {
  studentId: number;
  /** @minLength 1 */
  subject: string;
  /**
     * @minimum 0
     * @maximum 30
     */
  internalMarks: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  practicalMarks: number;
}

export interface SubjectSummary {
  subject: string;
  value: number;
}

export type StudentSummary = SubjectSummary;

export interface Dashboard {
  totalStudents: number;
  overallAttendance: number;
  averageMarks: number;
  recentStudents: Student[];
  subjectAttendance: SubjectSummary[];
  subjectPerformance: SubjectSummary[];
  lowAttendance: StudentSummary[];
}