# lib/api-zod/src/generated/types/attendance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceStatus } from './attendanceStatus';

export interface Attendance {
  id: number;
  studentId: number;
  subject: string;
  date: Date;
  status: AttendanceStatus;
}