# lib/api-zod/src/generated/types/attendanceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceInputStatus } from './attendanceInputStatus';

export interface AttendanceInput {
  studentId: number;
  /** @minLength 1 */
  subject: string;
  date: Date;
  status: AttendanceInputStatus;
}