# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Student } from './student';
import type { StudentSummary } from './studentSummary';
import type { SubjectSummary } from './subjectSummary';

export interface Dashboard {
  totalStudents: number;
  overallAttendance: number;
  averageMarks: number;
  recentStudents: Student[];
  subjectAttendance: SubjectSummary[];
  subjectPerformance: SubjectSummary[];
  lowAttendance: StudentSummary[];
}