# lib/api-zod/src/generated/types/performance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Performance {
  id: number;
  studentId: number;
  subject: string;
  internalMarks: number;
  practicalMarks: number;
}