# lib/api-zod/src/generated/types/performanceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PerformanceInput {
  studentId: number;
  /** @minLength 1 */
  subject: string;
  /**
     * @minimum 0
     * @maximum 30
     */
  internalMarks: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  practicalMarks: number;
}