# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  enrollmentNo: string;
  email: string;
  semester: number;
  branch: string;
}