# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  enrollmentNo: string;
  email: string;
  /** @minimum 1 */
  semester: number;
  /** @minLength 1 */
  branch: string;
}