# lib/api-zod/src/generated/types/studentSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubjectSummary } from './subjectSummary';

export type StudentSummary = SubjectSummary;